requirements/dev.txt
aiogram==2.25.1
black==23.11.0
flake8==5.0.4
python-dotenv==1.0.0